Rename dg_differentiate header to avoid conflict. Fix a couple of errors in unit test for GPU runs. Now unit test passes on cpu and gpu (gkylcas b5de8dc9e890e87f65659020596884489308766f).
